+++ packages/rum-core/src/domain/rumEventsCollection/action/actionCollection.ts
@@ -60,6 +60,8 @@ function processAction(
     ? {
         action: {
           id: action.id,
+          target: action.target,
+          position: action.position,
           loading_time: toServerDuration(action.duration),
           frustration: {
             type: action.frustrationTypes,