+++ packages/rum-core/src/rawRumEvent.types.ts
@@ -149,6 +149,13 @@ export interface RawRumActionEvent {
     resource?: Count
     target: {
       name: string
+      selector?: string
+      width?: number
+      height?: number
+    }
+    position?: {
+      x: number
+      y: number
     }
   }
   view?: {
@@ -187,6 +194,15 @@ export interface RumContext {
     type: string
     has_replay?: boolean
   }
+  view: {
+    id: string
+    referrer?: string
+    url: string
+    name?: string
+  }
+  action?: {
+    id: string | string[]
+  }
   synthetics?: {
     test_id: string
     result_id: string
@@ -204,26 +220,22 @@ export interface RumContext {
   }
 }
 
-export interface ViewContext extends Context {
+export interface ViewContext {
   service?: string
   version?: string
-  view: {
-    id: string
-    name?: string
-  }
+  id: string
+  name?: string
 }
 
-export interface ActionContext extends Context {
+export interface ActionContext {
   action: {
     id: string | string[]
   }
 }
 
-export interface UrlContext extends Context {
-  view: {
-    url: string
-    referrer: string
-  }
+export interface UrlContext {
+  url: string
+  referrer: string
 }
 
 export interface InternalContext {